backend/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.1.0
pydantic>=2.7.0
pdfplumber>=0.11.0
pandas>=2.2.0
openpyxl>=3.1.5
python-dotenv>=1.0.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
